Re-order ca-cert, proxy and repo config steps
This eliminates some issues where repos couldn't be refreshed due to
missing certificates or proxy configuration.

In some cases not all packages required for certificate handling are
available in the base install. Due to this there are two certificate
refreshes, one before, one after package installation. Only the second
one is counted as play aborting failure.

If custom certificates are required for package installation all
packages required for certificate handling need to be in the base
installation.