Add WSL provisioning
This expects a locally available appx installer. If not included in the
image, make sure it gets downloaded before this role gets executed.

Basic provisoning (setup of management user and ssh configuration) is
performed, so a distribution can be configured through ansible, if
required. Those steps always are executed, but the sshd is only started
if sshd_port is configured, and sshd_service is set to true. This will
also configure the Windows firewall to allow connections to that port.

The SSH settings are per WSL instance. A sample WSL confguration looks
like this:

wsl_distributions:
  openSUSE-Leap-15-1:
    installer: 'c:\ci\opensuse.appx'
    installer_type: appx
    location: 'd:\suse'
    sshd_port: 8022
    sshd_service: True