Allow dropping non-oss for SuSE through suse_core_repos
suse_core_repos is a list, containing 'oss' and 'non-oss' per
default. To disable for example oss, define it with just 'oss' as
entry.

Also, allow setting suse_update_url_prefix for non-standard update
URLs, as required for example for ports. Note that update repos are
only generated for Leap.