Introduce repo_ssl_verify variable to control certificate checking
So far implemented for suse only. Freshly bootstrapped containers tend
to have missing certificate files, which can't be installed as the
repository servers SSL certificate can't be verified without it.

Ideally we'd be forcing certificates in instead, but that's not always
practical, and especially in cases where we control the mirror and
don't leave our own network temporarily disabling this is not an
issue.

The parameter should be passed as vars argument in an include_role,
and not be set as host/group variables - the initial provisioning step
pulls in all relevant packages, and the next proper run overrides this.